+++ packages/spacecat-shared-utils/test/url-helpers.test.js
@@ -1,0 +1,99 @@
+/*
+ * Copyright 2024 Adobe. All rights reserved.
+ * This file is licensed to you under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License. You may obtain a copy
+ * of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under
+ * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+ * OF ANY KIND, either express or implied. See the License for the specific language
+ * governing permissions and limitations under the License.
+ */
+
+/* eslint-env mocha */
+
+import { expect } from 'chai';
+import {
+  composeBaseURL,
+  prependSchema,
+  stripPort,
+  stripTrailingDot,
+  stripTrailingSlash,
+  stripWWW,
+} from '../src/url-helpers.js';
+
+describe('URL Utility Functions', () => {
+  describe('prependSchema', () => {
+    it('should prepend "https://" schema if not present', () => {
+      expect(prependSchema('example.com')).to.equal('https://example.com');
+    });
+
+    it('should not modify the URL if schema is already present', () => {
+      expect(prependSchema('https://example.com')).to.equal('https://example.com');
+      expect(prependSchema('http://example.com')).to.equal('http://example.com');
+    });
+  });
+
+  describe('stripPort', () => {
+    it('should remove port number from the end of the URL', () => {
+      expect(stripPort('example.com:80')).to.equal('example.com');
+      expect(stripPort('example.com:8080')).to.equal('example.com');
+    });
+
+    it('should not modify the URL if no port number present', () => {
+      expect(stripPort('example.com')).to.equal('example.com');
+    });
+  });
+
+  describe('stripTrailingDot', () => {
+    it('should remove trailing dot from the end of the URL', () => {
+      expect(stripTrailingDot('example.com.')).to.equal('example.com');
+      expect(stripTrailingDot('.example.com')).to.equal('.example.com');
+      expect(stripTrailingDot('example.com.abc')).to.equal('example.com.abc');
+    });
+
+    it('should not modify the URL if no trailing dot present', () => {
+      expect(stripTrailingDot('example.com')).to.equal('example.com');
+    });
+  });
+
+  describe('stripTrailingSlash', () => {
+    it('should remove trailing slash from the end of the URL', () => {
+      expect(stripTrailingSlash('example.com/')).to.equal('example.com');
+      expect(stripTrailingSlash('example.com//')).to.equal('example.com/');
+      expect(stripTrailingSlash('/example.com/')).to.equal('/example.com');
+      expect(stripTrailingSlash('example.com/abc/')).to.equal('example.com/abc');
+    });
+
+    it('should not modify the URL if no trailing slash present', () => {
+      expect(stripTrailingSlash('example.com')).to.equal('example.com');
+    });
+  });
+
+  describe('stripWWW', () => {
+    it('should remove "www." from the beginning of the URL', () => {
+      expect(stripWWW('www.example.com')).to.equal('example.com');
+    });
+
+    it('should not modify the URL if "www." is not present', () => {
+      expect(stripWWW('example.com')).to.equal('example.com');
+    });
+
+    it('should remove "www." even if schema is present', () => {
+      expect(stripWWW('https://www.example.com')).to.equal('https://example.com');
+      expect(stripWWW('http://www.example.com')).to.equal('http://example.com');
+    });
+  });
+
+  describe('composeBaseURL', () => {
+    it('should compose base URL by applying all transformations', () => {
+      expect(composeBaseURL('https://www.example.com:8080/')).to.equal('https://example.com');
+      expect(composeBaseURL('http://www.example.com:8080/')).to.equal('http://example.com');
+      expect(composeBaseURL('www.example.com:8080/')).to.equal('https://example.com');
+      expect(composeBaseURL('example.com:8080/')).to.equal('https://example.com');
+      expect(composeBaseURL('https://example.com/')).to.equal('https://example.com');
+      expect(composeBaseURL('http://example.com/')).to.equal('http://example.com');
+      expect(composeBaseURL('example.com/')).to.equal('https://example.com');
+    });
+  });
+});